+++ stroom-dashboard/stroom-dashboard-client/src/main/java/stroom/dashboard/client/table/FormatPresenter.java
@@ -85,8 +85,6 @@ public class FormatPresenter extends MyPresenterWidget<FormatPresenter.FormatVie
             final Format format = getFormat();
             if (!EqualsUtil.isEquals(format, field.getFormat())) {
                 fieldChangeConsumer.accept(field, field.copy().format(format).build());
-                tablePresenter.setDirty(true);
-                tablePresenter.clearAndRefresh();
             }
         }